@@ -1,2 +1,1 @@
-import './styles/app.css';
-const $ = require('jquery');
+console.log(44)
